Removed support for .NET Framework-based ASP .NET Core applications; Re-architected code base to use more relevant class names and directory structure; Now supports .NET 6, 7, and 8; Updated readme and changelog
