+++ OwaspHeaders.Core.Example/Controllers/HomeController.cs
@@ -1,0 +1,22 @@
+﻿using Microsoft.AspNetCore.Mvc;
+
+namespace OwaspHeaders.Core.Example.Controllers
+{
+    [ApiController]
+    [Route("/")]
+    public class HomeController : ControllerBase
+    {
+        private readonly ILogger<HomeController> _logger;
+
+        public HomeController(ILogger<HomeController> logger)
+        {
+            _logger = logger;
+        }
+
+        [HttpGet(Name = "/")]
+        public IEnumerable<string> Get()
+        {
+            return HttpContext.Response.Headers.Select(h => h.ToString()).ToArray();
+        }
+    }
+}
