+++ OwaspHeaders.Core.Example/Properties/launchSettings.json
