+++ OwaspHeaders.Core.Example/appsettings.Development.json
