+++ OwaspHeaders.Core.Example/appsettings.json
