+++ OwaspHeaders.Core.Tests/CustomHeaders/XRemovePoweredByOptions.cs
@@ -1,0 +1,46 @@
+﻿using System.Threading.Tasks;
+using OwaspHeaders.Core;
+using OwaspHeaders.Core.Extensions;
+using Xunit;
+
+namespace OwaspHeaders.Core.Tests.CustomHeaders
+{
+    public class XRemovePoweredByOptions : SecureHeadersTests
+    {
+        [Fact]
+        public async Task When_RemovePoweredByHeaderCalled_Header_Is_Present()
+        {
+            // arrange
+            var headerPresentConfig = SecureHeadersMiddlewareBuilder.CreateBuilder()
+                .RemovePoweredByHeader().Build();
+            var secureHeadersMiddleware = new SecureHeadersMiddleware(_onNext, headerPresentConfig);
+
+            // act
+            await secureHeadersMiddleware.InvokeAsync(_context);
+
+            // assert
+            Assert.True(headerPresentConfig.RemoveXPoweredByHeader);
+            Assert.False(_context.Response.Headers.ContainsKey(Constants.PoweredByHeaderName));
+            Assert.False(_context.Response.Headers.ContainsKey(Constants.ServerHeaderName));
+        }
+
+        [Fact]
+        public async Task When_RemovePoweredByHeaderNotCalled_Header_Not_Present()
+        {
+            // arrange
+            var headerPresentConfig = SecureHeadersMiddlewareBuilder.CreateBuilder()
+                .Build();
+            var secureHeadersMiddleware = new SecureHeadersMiddleware(_onNext, headerPresentConfig);
+
+            // act
+            await secureHeadersMiddleware.InvokeAsync(_context);
+
+            // assert
+            Assert.False(headerPresentConfig.RemoveXPoweredByHeader);
+            // Am currently running the 2.1.300 Preview 1 build of the SDK
+            // and the server doesn't seem to add this header.
+            // Therefore this assert is commented out, as it will always fail
+            // Assert.True(_context.Response.Headers.ContainsKey(Constants.PoweredByHeaderName));
+        }
+    }
+}
