+++ src/SecureHeadersMiddleware.cs
